docs(ref): Find a place to comment on --cap-lints
Looking over the points of confusion highlighted in #5998,
one is in rustc which is agnostic of any of this policy and the rest are
in historical documents.

Inspired by previous comments, I figured we could fit this into a
discussion of `[lints]` by talking about the scope of the feature.

Fixes #5998