chore: tracing be compat with rustc_log
The current version of rustfix cannot be used in rust-lang/rust due
to the dependency on `tracing`. `tracing-core` is currently locked
to 0.1.30, which is too old for the version in the cargo repo
(it looks like they are trying to avoid a known deadlock).

See: https://rust-lang.zulipchat.com/#narrow/stream/246057-t-cargo/topic/rustfix.20is.20incompatible.20with.20rust-lang.2Frust