chore(ci): cargo-hack `--no-dev-deps`
One thing to note is that the use of the --all-targets flag (that builds
tests/examples) needed to be removed and the --no-dev-deps flag is needed.
This is because the dev-dependencies use versionless manifests that cannot
be read by the old compiler.

That said, we are not running tests with the old compiler, so I don't see
a problem if the test suite cannot be built on the old compiler.

Co-authored-by: Taiki Endo <te316e89@gmail.com>