Explore location to generate SBOM precursor files
Similar to the generation of `depinfo` files, a function is called to
generated SBOM precursor file named `output_sbom`. It takes the
`BuildRunner` & the current `Unit`. The `sbom` flag can be specified as
a cargo build option, but it's currently not configured correctly. To
test the generation the flag is set to `true`.

This passes in the cargo build config `sbom`.