test: deal with lockfile checksums dynamically
When generating local registries in the testsuite, the specific package
checksums may vary from the original simply due to different `zlib`
versions. For example, Fedora 40 is now using `zlib-ng-compat`. We can
account for this dynamically by either formatting the checksum in the
expected output string, or using `generate-lockfile` on projects and
their expected output.