feat: delay file patches register to next round of resolution
Only after a few dependency resolution happen, can a `[patch]` entry
with patch files know which version it is going to patch. Hence we need
to "delay" the patch process until at least one resolution is done, and
re-resolves the graph is any patch become applicable with the newly
resolved dependency graph.