Add more high level traces
This accounts for more time when running rustc (which turns out to be a
significant amount of time).

I'm less sure about the start/wait calls but I'm seeing very different
results from different builds of `cargo` and some have some large
amounts of unaccounted time that I want to dig into (and callgrind and
and samply haven't helped).