fix(toml): Warn when edition is unuset, even when MSRV is unset
This came up in #14108.  This got overlooked when we added the MSRV
limits to the warning.

Users can silience this by setting a very old MSRV.
I didn't bother finding a way to word a specialized message for this.
Wording for this case seemed hard and I figure someone in this situation
knows how to resolve it.
Instead, we are targeting the majority of users who aren't setting a
`package.rust-version` in the first place.