Add a JSON output mode to `cargo package`.
In multi-crate workspaces, it can be ambiguous where files are sourced
from when building packages. This extends the existing `--list` command
line flag to optionally emit JSON when called as `-l=json` or
`--list=json`. Existing behaviour is unaffected.

The JSON format is slightly different to that proposed in #11666: it
exposes the distinction within `cargo::ops::cargo_package` itself around
whether an archive file is generated or sourced from the disk. A new
crate created with `cargo new --lib foo` generates this output:

```json
{
  ".cargo_vcs_info.json": {
    "Generated": {
      "VcsInfo": {
        "git": {
          "sha1": "af0bf5917934ec4684cd23763f90c16242d7c5ce"
        },
        "path_in_vcs": ""
      }
    }
  },
  ".gitignore": {
    "OnDisk": "/tmp/foo/.gitignore"
  },
  "Cargo.toml": {
    "Generated": "Manifest"
  },
  "Cargo.toml.orig": {
    "OnDisk": "/tmp/foo/Cargo.toml"
  },
  "src/lib.rs": {
    "OnDisk": "/tmp/foo/src/lib.rs"
  }
}
```

I'm open minded as to whether the `VcsInfo` should actually be included,
but it feels like there's little harm in doing so.

Fixes #11666, and is at least extremely relevant to #13953.