introduce no-path-check for install command
Provides the option to allow a user to trigger an install request that
will not generate a warning if an installed program is not available in
the `PATH`. This can be helpful when the install root is configured
to an interim location (e.g. cross-compiling) and there is a desire to
suppress any path checks warnings not relevant for the install case.

Signed-off-by: James Knight <james.d.knight@live.com>